+++ RELEASENOTES.md
@@ -24,6 +24,10 @@
         video track are available. The default value is `false` which means
         selecting a video track is prioritized.
 *   Extractors:
+    *   Add additional AV1C parsing to MP4 extractor to retrieve
+        `ColorInfo.colorSpace`, `ColorInfo.colorTransfer`, and
+        `ColorInfo.colorRange` values
+        ([#692](https://github.com/androidx/media/pull/692)).
 *   Audio:
 *   Video:
     *   Add workaround for a device issue on Galaxy Tab S7 FE, Chromecast with